+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project follows [Semantic Versioning](http://semver.org/).
 - Fixed the incorrect input field size for autocomplete when the list of
   available search parameters contains only one value.
 - Fixed the incorrect positioning of the "connect with SMART on FHIR" checkbox.
+- Excluded resource types without visible search parameters from the cohort
+  criteria.
 
 ## [7.9.2] - 2026-04-10
 ### Changed